requirements.txt
git+https://github.com/Shriyansh92/yt-dl@master
pyrogram==1.4.16
TgCrypto
py-tgcalls==0.8.6
python-dotenv
youtube_search_python
requests
aiohttp
aiofiles
asyncio
python-arq
youtube_search
search_engine_parser
ffmpeg-python
ffmpeg
Pillow
ujson
psutil
telegram
python-telegram-bot==13.11
